+++ sdk/core/core-client-rest/src/getClient.ts
@@ -3,18 +3,19 @@
 
 import { KeyCredential, TokenCredential } from "@azure/core-auth";
 import { createDefaultPipeline } from "./clientHelpers";
-import { ClientOptions, HttpResponse, Client } from "./common";
+import { ClientOptions, HttpResponse, ClientResource, StreamableMethod, Client } from "./common";
 import { RequestParameters } from "./pathClientTypes";
-import { buildSendRequest, isCredential } from "./helpers/getClientHelpers";
+import { buildSendRequest, isCredential, StreamType } from "./helpers/getClientHelpers";
+import { HttpNodeStreamResponse } from ".";
 
 /**
- * Creates a client with a default pipeline
+ * Creates a client with a default pipeline that has methods to get response as Streams
  * @param baseUrl - Base endpoint for the client
  * @param options - Client options
  */
 export function getClient(baseUrl: string, options?: ClientOptions): Client;
 /**
- * Creates a client with a default pipeline
+ * Creates a client with a default pipeline that has methods to get response as Streams
  * @param baseUrl - Base endpoint for the client
  * @param credentials - Credentials to authenticate the requests
  * @param options - Client options
@@ -40,102 +41,147 @@ export function getClient(
 
   const pipeline = createDefaultPipeline(baseUrl, credentials, clientOptions);
   const { allowInsecureConnection } = clientOptions;
-  const client = (path: string, ...args: Array<any>) => {
-    return {
-      get: (options: RequestParameters = {}): Promise<HttpResponse> => {
-        return buildSendRequest(
+
+  const resourceWithAsStream = (
+    path: string,
+    ...args: Array<any>
+  ): ClientResource<StreamableMethod> => ({
+    get(options: RequestParameters = {}) {
+      const operation = (streamType?: StreamType) =>
+        buildSendRequest(
           "GET",
           clientOptions,
           baseUrl,
           path,
           pipeline,
+          streamType,
           { allowInsecureConnection, ...options },
           args
         );
-      },
-      post: (options: RequestParameters = {}): Promise<HttpResponse> => {
-        return buildSendRequest(
+
+      return getStreamableMethods(operation);
+    },
+    post(options: RequestParameters = {}) {
+      const operation = (streamType?: StreamType) =>
+        buildSendRequest(
           "POST",
           clientOptions,
           baseUrl,
           path,
           pipeline,
+          streamType,
           { allowInsecureConnection, ...options },
           args
         );
-      },
-      put: (options: RequestParameters = {}): Promise<HttpResponse> => {
-        return buildSendRequest(
+      return getStreamableMethods(operation);
+    },
+    put: (options: RequestParameters = {}) => {
+      const operation = (streamType?: StreamType) =>
+        buildSendRequest(
           "PUT",
           clientOptions,
           baseUrl,
           path,
           pipeline,
+          streamType,
           { allowInsecureConnection, ...options },
           args
         );
-      },
-      patch: (options: RequestParameters = {}): Promise<HttpResponse> => {
-        return buildSendRequest(
+      return getStreamableMethods(operation);
+    },
+    patch: (options: RequestParameters = {}) => {
+      const operation = (streamType?: StreamType) =>
+        buildSendRequest(
           "PATCH",
           clientOptions,
           baseUrl,
           path,
           pipeline,
+          streamType,
           { allowInsecureConnection, ...options },
           args
         );
-      },
-      delete: (options: RequestParameters = {}): Promise<HttpResponse> => {
-        return buildSendRequest(
+      return getStreamableMethods(operation);
+    },
+    delete: (options: RequestParameters = {}) => {
+      const operation = (streamType?: StreamType) =>
+        buildSendRequest(
           "DELETE",
           clientOptions,
           baseUrl,
           path,
           pipeline,
+          streamType,
           { allowInsecureConnection, ...options },
           args
         );
-      },
-      head: (options: RequestParameters = {}): Promise<HttpResponse> => {
-        return buildSendRequest(
+
+      return getStreamableMethods(operation);
+    },
+    head: (options: RequestParameters = {}) => {
+      const operation = (streamType?: StreamType) =>
+        buildSendRequest(
           "HEAD",
           clientOptions,
           baseUrl,
           path,
           pipeline,
+          streamType,
           { allowInsecureConnection, ...options },
           args
         );
-      },
-      options: (options: RequestParameters = {}): Promise<HttpResponse> => {
-        return buildSendRequest(
+      return getStreamableMethods(operation);
+    },
+    options: (options: RequestParameters = {}) => {
+      const operation = (streamType?: StreamType) =>
+        buildSendRequest(
           "OPTIONS",
           clientOptions,
           baseUrl,
           path,
           pipeline,
+          streamType,
           { allowInsecureConnection, ...options },
           args
         );
-      },
-      trace: (options: RequestParameters = {}): Promise<HttpResponse> => {
-        return buildSendRequest(
+      return getStreamableMethods(operation);
+    },
+    trace: (options: RequestParameters = {}) => {
+      const operation = (streamType?: StreamType) =>
+        buildSendRequest(
           "TRACE",
           clientOptions,
           baseUrl,
           path,
           pipeline,
+          streamType,
           { allowInsecureConnection, ...options },
           args
         );
-      },
-    };
-  };
+      return getStreamableMethods(operation);
+    },
+  });
 
   return {
-    path: client,
-    pathUnchecked: client,
+    path: resourceWithAsStream,
+    pathUnchecked: resourceWithAsStream,
     pipeline,
+  };
+}
+
+interface SendRequestOperation {
+  (streamType: "NodeJS"): Promise<HttpNodeStreamResponse>;
+  (streamType?: StreamType): Promise<HttpResponse>;
+}
+
+function getStreamableMethods(sendRequest: SendRequestOperation): StreamableMethod {
+  return {
+    then: async function (onfulfilled) {
+      const result = await sendRequest();
+      return onfulfilled(result);
+    },
+    async asNodeStream() {
+      return sendRequest("NodeJS");
+    },
   };
 }
